Added the ErrorController and moved error handling components of the HomeController here
